\name{initialHy}
\alias{initialHy}
\title{initialHy}
\usage{
ih <- initialHy(df = df, Atgt = Atgt, Acmp = Acmp,
                Atgt_dist = Atgt_dist, Acmp_dist = Acmp_dist,
                Atgt_cl = Atgt_cl, Acmp_cl = Acmp_cl,
                Actx_items = Actx_items)
}
\description{
Sets up the initial hypothesis.
}
\examples{
ih <- initialHy(df = df, Atgt = Atgt, Acmp = Acmp,
                Atgt_dist = Atgt_dist, Acmp_dist = Acmp_dist,
                Atgt_cl = Atgt_cl, Acmp_cl = Acmp_cl,
                Actx_items = Actx_items)
}
